# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Xatkit-SPL API",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "https://xatkit-spl-backend.onrender.com"
    }
  ],
  "paths": {
...
    "/api/v1/chatbots": {
      "post": {
        "tags": [
          "Chatbots"
        ],
        "summary": "Create a chatbot",
        "description": "Create a chatbot",
        "operationId": "createAChatbot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "example": "newChatbot"
                  },
                  "intents": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": [
                        "63debe9878d39a8cc462f7b9",
                        "63debe9878d39a8cc462fa29"
                      ]
                    },
                    "example": [
                      [
                        "63debe9878d39a8cc462f7b9",
                        "63debe9878d39a8cc462fa29"
                      ]
                    ]
                  },
                  "name": {
                    "type": "string",
                    "example": "newChatbot"
                  }
                }
              },
              "example": {
                "description": "Create a chatbot",
                "intents": [
                  "63debe9878d39a8cc462f7b9",
                  "63debe9878d39a8cc462fa29"
                ],
                "name": "newChatbot"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/chatbots/mine/{chatbot_id}": {
      "parameters": [
        {
          "name": "chatbot_id",
          "in": "path",
          "description": "ID of the chatbot to retrieve",
          "required": true,
          "schema": {
            "type": "string",
            "example": "63debede78d39a8cc462f7c2"
          }
        }
      ],
      "get": {
        "tags": [
          "Chatbots"
        ],
        "summary": "Get my chatbot by ID",
        "description": "Get my chatbot by ID",
        "operationId": "getMyChatbotById",
        "parameters": [
          {
            "name": "chatbot_id",
            "in": "path",
            "description": "ID of the chatbot to get",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "Chatbots"
        ],
        "summary": "Update my chatbot by ID",
        "description": "Update my chatbot by ID",
        "operationId": "updateMyChatbotById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "intents": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "63debe9878d39a8cc462f7b9"
                    },
                    "example": [
                      "63debe9878d39a8cc462f7b9",
                      "63debe9f78d39a8cc462f7bb"
                    ]
                  }
                }
              },
              "example": {
                "intents": [
                  "63debe9878d39a8cc462f7b9",
                  "63debe9f78d39a8cc462f7bb"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Chatbots"
        ],
        "summary": "Delete my chatbot by ID",
        "description": "Delete my chatbot by ID",
        "operationId": "deleteMyChatbotById",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
}